refactor: move envs import to infrastructure config directory
Update import paths for envs configuration to use the new location in infrastructure layer
Update test mocks and setup to reflect the new envs location and simplify test configurations
