+++ src/config/tests/__mocks__/env.mock.ts
@@ -5,7 +5,7 @@ const mockEnv = {
   SUPABASE_ANON_KEY: "test-anon-key",
   JWT_SECRET: "jwt-test-secret-with-at-least-32-characters-long",
   DATABASE_URL: "postgresql://test:test@localhost:5432/test",
-  PASSPHRASE: "passphrase-for-encryption",
+  PASSPHRASE: "test-passphrase",
 };
 
 Object.freeze(mockEnv);
