+++ src/config/tests/example.test.ts
@@ -25,8 +25,8 @@ describe("Jest Configuration", () => {
 
   it("should be able to use test environment variables", () => {
     process.env.SUPABASE_URL = "https://test.supabase.co";
-    expect(process.env.NODE_ENV).toBe("test");
-    expect(process.env.PORT).toBe("3001");
+    expect(process.env.NODE_ENV).toBe("dev");
+    expect(process.env.PORT).toBe("3008");
     expect(process.env.SUPABASE_URL).toBe("https://test.supabase.co");
   });
